+++ mocktail-dashboard/src/utils/paths.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-undef */
 const BASE_URL = process.env.NODE_ENV === 'development' ? 'http://localhost:4000' : '';
 
-const API_CORE_URL = `${BASE_URL}/core/v1`;
+export const API_CORE_URL = `${BASE_URL}/core/v1`;
 
 export const ALL_APIS = `${API_CORE_URL}/apis`;
 export const SAVE_API = `${API_CORE_URL}/api`;